AnaTutorial Updates
I updated some of the examples to match changes made in our ECCE macros repo. This is decalring the different far forward detector configurations for IP6 and IP8 which are now distinct functions. I also removed the string declared in Jet_FwdEval which is no longer present in the common macro (its now a directory so things got a little odd)

To-do: In AnaTutorial.cc we look for a track map named "SvtxTrackMap" while the default ECCe map is now "TrackMap". This doesn't cause a crash but this part of the tutorial doesn't run for ECCE. We can remove "Svtx" to fix this but it will cause a divergence from sPHENIX tutorials.